requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-header
